ajout d'une page pour gérer les différents cache + ajout d'une page administration

+++ app/src/Utils/FileUtils.php
@@ -52,7 +52,10 @@ class FileUtils {
         }
 
         closedir($dir);
-        rmdir($path);
+
+        if ($result) {
+            rmdir($path);
+        }
 
         return $result;
     }
